ref: move native file to ts
