+++ src/libs/cropOrRotateImage/index.native.ts
@@ -1,27 +1,18 @@
 import RNImageManipulator from '@oguzhnatly/react-native-image-manipulator';
 import RNFetchBlob from 'react-native-blob-util';
+import {CropOrRotateImage} from './types';
 
 /**
  * Crops and rotates the image on ios/android
- *
- * @param {String} uri
- * @param {Array<Object>} actions
- * @param {Object} options
- * @returns {Promise<Object>} Returns cropped and rotated image
  */
-function cropOrRotateImage(uri, actions, options = {}) {
-    return new Promise((resolve) => {
+const cropOrRotateImage: CropOrRotateImage = (uri, actions, options) =>
+    new Promise((resolve) => {
         RNImageManipulator.manipulate(uri, actions, options).then((result) => {
             RNFetchBlob.fs.stat(result.uri.replace('file://', '')).then(({size}) => {
-                resolve({
-                    ...result,
-                    size,
-                    type: options.type || 'image/jpeg',
-                    name: options.name || 'fileName.jpg',
-                });
+                const file = Object.assign(result, {size, type: options.type || 'image/jpeg', name: options.name || 'fileName.jpg'});
+                resolve(file);
             });
         });
     });
-}
 
 export default cropOrRotateImage;
